Fingerprint field for contacts, not filled yet
The regular Fingerprint doesn't work here because it requires the salt, which is
not part of the public Destination. The salt is published with the Destination
in the public address book, and could be conveyed between Bote apps via NFC or
QR code, but that doesn't work for Destinations published in other spaces.